Update core to 3cdd461 (Feb 28)

+++ deps/chakrashim/core/lib/Runtime/Runtime.h
@@ -262,7 +262,7 @@ namespace Js
     // asm.js
     namespace ArrayBufferView
     {
-        enum ViewType: int;
+        enum ViewType: uint8;
     }
     struct EmitExpressionInfo;
     struct AsmJsModuleMemory;
@@ -444,13 +444,13 @@ enum tagDEBUG_EVENT_INFO_TYPE
 #include "Types/JavascriptStaticEnumerator.h"
 #include "Library/ExternalLibraryBase.h"
 #include "Library/JavascriptLibraryBase.h"
+#include "Library/MathLibrary.h"
 #include "Base/ThreadContextInfo.h"
 #include "Library/JavascriptLibrary.h"
 
 #include "Language/JavascriptExceptionOperators.h"
 #include "Language/JavascriptOperators.h"
 
-#include "Library/MathLibrary.h"
 #include "Library/WasmLibrary.h"
 // xplat-todo: We should get rid of this altogether and move the functionality it
 // encapsulates to the Platform Agnostic Interface
